#989 toggles favorite instead of giving body
