+++ src/backend/src/routes/projects.routes.ts
@@ -46,5 +46,6 @@ projectRouter.post(
 );
 projectRouter.post('/:wbsNum/set-team', nonEmptyString(body('teamId')), validateInputs, ProjectsController.setProjectTeam);
 projectRouter.delete('/:wbsNum/delete', ProjectsController.deleteProject);
+projectRouter.post('/:projectId/favorite', ProjectsController.toggleFavorite);
 
 export default projectRouter;
