+++ src/backend/tests/users.test.ts
@@ -92,51 +92,4 @@ describe('Users', () => {
       expect(res.slackId).toStrictEqual('slack');
     });
   });
-
-  describe('favoriteProjects', () => {
-    test('fails when project does not exist', async () => {
-      jest.spyOn(prisma.project, 'findUnique').mockResolvedValue(null);
-      const fakeProjectId = 100000;
-      await expect(() => UsersService.updateUserFavorites(fakeProjectId, batman, true)).rejects.toThrow(
-        new NotFoundException('Project', fakeProjectId)
-      );
-      expect(prisma.project.findUnique).toBeCalledTimes(1);
-    });
-    test('fails when user does not exist', async () => {
-      jest.spyOn(prisma.project, 'findUnique').mockResolvedValue(prismaProject1);
-      jest.spyOn(prisma.user, 'findFirst').mockResolvedValue(null);
-
-      const fakeUserId = 100000;
-      await expect(() =>
-        UsersService.updateUserFavorites(prismaProject1.projectId, { ...batman, userId: fakeUserId }, true)
-      ).rejects.toThrow(new NotFoundException('User', fakeUserId));
-      expect(prisma.project.findUnique).toBeCalledTimes(1);
-      expect(prisma.user.findFirst).toBeCalledTimes(1);
-    });
-    test('completes succesfully when adding', async () => {
-      jest.spyOn(prisma.project, 'findUnique').mockResolvedValue(prismaProject1);
-      jest.spyOn(prisma.user, 'findFirst').mockResolvedValue(batman);
-      jest.spyOn(prisma.user, 'update').mockResolvedValue(batman);
-
-      const res = await UsersService.updateUserFavorites(prismaProject1.projectId, batman, true);
-
-      expect(res).toBe(prismaProject1);
-      expect(prisma.project.findUnique).toBeCalledTimes(1);
-      expect(prisma.user.findFirst).toBeCalledTimes(1);
-      expect(prisma.user.update).toBeCalledTimes(1);
-    });
-
-    test('completes succesfully when deleting', async () => {
-      jest.spyOn(prisma.project, 'findUnique').mockResolvedValue(prismaProject1);
-      jest.spyOn(prisma.user, 'findFirst').mockResolvedValue(batman);
-      jest.spyOn(prisma.user, 'update').mockResolvedValue(batman);
-
-      const res = await UsersService.updateUserFavorites(prismaProject1.projectId, batman, false);
-
-      expect(res).toBe(prismaProject1);
-      expect(prisma.project.findUnique).toBeCalledTimes(1);
-      expect(prisma.user.findFirst).toBeCalledTimes(1);
-      expect(prisma.user.update).toBeCalledTimes(1);
-    });
-  });
 });
